+++ src/frontend/servers/y-provider/jest.config.js
@@ -6,6 +6,7 @@ var config = {
   },
   moduleNameMapper: {
     '^@/(.*)$': '<rootDir>/../src/$1',
+    '^@blocknote/server-util$': '<rootDir>/../__mocks__/mock.js',
   },
 };
 export default config;